Initial work on species containment stacked bar graph
